+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added `ContextualColor` model and its documentation.
 * Added additional options to `SearchTextField` to allow for custom icons and colors. Improved overall look and spacing of the search text field.
 * Moved `shapeBorder` from `DropDown` class to `DropDownOptions` as `border`. Changed default to a rounded border of `24.0` instead of `15.0`.
+* Added new constructors and static helper methods to `SelectedListItem` class.
 
 ## 0.0.1